Moved some tests to it/ folder as are really integration tests
